Allow hash of arrays of hashes for the config
I've created a way to push configs via a hash of arrays of hashes. Attached in a library to handle merging the existing attributes with the new attributes. I don't have documentation on using it, but the format only requires a bit of complex config translation and the user can just read the commit.