Support the Go MyFace port and the YAML config
This installs the GO port of MyFace and the YAML config that is used in the port. The existing attribute system doesn't work as the config is different so I included a new set of attributes for use with this port. The packages are also different as are the config paths so there's logic to choose the right names.